requirements.txt
websocket-client
pandas
gradio